fixed elevated priv issue fixed multiple alerts

+++ Application/monitor_and_block.py
@@ -17,20 +17,28 @@ def check_for_processes(process_name):
 
 
 # Terminate the Process
-def block_process(process):
+def block_process(process, warned_processes):
 
     try:
-        process.terminate()
-        # Allow the process to terminate
-        process.wait()
-        show_warning()
+        if process not in warned_processes:
+            process.terminate()
+            # Allow the process to terminate
+            process.wait()
+            show_warning()
+            warned_processes.add(process.info['name'].lower())
     except psutil.NoSuchProcess:
         # In case the process has already been terminated
         pass
-
+    except psutil.AccessDenied:
+        # Handle the AccessDenied exception
+        print(f"Access denied for terminating process: {process.info['name']} (pid: {process.pid}).")
+        print(f"ScamWatch needs elevated")
 
 # Continiously monitor process list
 def monitor_process():
+
+    # Set to keep track of warned pids
+    warned_processes = set()
 
     while True:
         if is_paused():
@@ -42,8 +50,9 @@ def monitor_process():
             for process_name in EXE_NAME_LIST:
                 processes = check_for_processes(process_name)
                 for process in processes:
-                    block_process(process)
+                    block_process(process, warned_processes)
             time.sleep(1)
+            warned_processes.clear() # Clear set each cycle to allow for new warnings
 
 # allow a short connection by pausing monitoring
 def one_time_connection():
@@ -53,11 +62,9 @@ def one_time_connection():
 def show_warning():
     ctypes.windll.user32.MessageBoxW(
         0,
-        "WARNING! This may be a remote connection scam!\nDon't worry, however, we have blocked this program and you are safe!",
+        "WARNING! This may be a remote connection scam!\nDon't worry, however, we have blocked and you are safe!",
         "ScamWatch Alert",
         0x40 | 0x1 | 0x40000  # MB_ICONWARNING | MB_OK | MB_TOPMOST
     )
-
-show_warning()
